+++ src/calendar/day/period/index.tsx
@@ -4,13 +4,13 @@ import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import {TouchableWithoutFeedback, Text, View, ViewStyle} from 'react-native';
 
-// @ts-expect-error
-import {shouldUpdate} from '../../../component-updater';
+import {Theme, DayState} from '../../../types';
+import {shouldUpdate} from '../../../componentUpdater';
 import * as defaultStyle from '../../../style';
 import styleConstructor from './style';
 import Dot from '../dot';
 import {MarkingProps} from '../marking';
-import {Theme, DayState} from '../../../types';
+
 
 interface PeriodDayProps {
   state?: DayState;